% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalsoundr-package.R
\docType{package}
\name{animalsoundr-package}
\alias{animalsoundr}
\alias{animalsoundr-package}
\title{animalsoundr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gguyver/animalsoundr}
  \item \url{https://gguyver.github.io/animalsoundr/}
  \item Report bugs at \url{https://github.com/gguyver/animalsoundr/issues}
}

}
\author{
\strong{Maintainer}: Glen Guyver-Fletcher \email{glen.guyver-fletcher@warwick.ac.uk}

}
\keyword{internal}
