% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{farm_animals}
\alias{farm_animals}
\title{Farm animals with associated sounds}
\format{
A data frame with 2 columns and 5 rows.
\describe{
\item{animal}{The farm animal as a string.}
\item{sound}{The sound the farm animal makes, as a string.}
}
}
\usage{
farm_animals
}
\description{
A dataset listing some common farm animals and the sound that they
stereotypically make
}
\keyword{datasets}
