% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Print an animal with provided sound}
\usage{
animal_sounds(animal, sound = NULL)
}
\arguments{
\item{animal}{An animal name, e.g. "Tiger", "Dog"}

\item{sound}{A sound the animal will make, e.g. "Roar", "Woof", which is pasted together using \code{\link[=paste]{paste()}}.}
}
\description{
This function takes two strings and returns a \strong{pasted output}.
}
\examples{
animal_sounds("dog", "woof")
}
