% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_sounds.R
\name{animal_sounds}
\alias{animal_sounds}
\title{Make animal sounds}
\usage{
animal_sounds(animal, sound)
}
\arguments{
\item{animal}{A character string of animal name}

\item{sound}{A character string of animal sound}
}
\value{
A character string with a sentence
}
\description{
Print the given sound of a given animal.
}
\examples{

animal_sounds(animal = "cat", sound = "meow")

}
